fix: allow users to confirm navigation with enter/escape buttons
Coses #1271
