+++ src/components/sandbox/components/AddNewVm.tsx
@@ -31,6 +31,7 @@ import { createVirtualMachine, getVmName, getVirtualMachineCost } from '../../..
 import { SandboxObj, DropdownObj, SizeObj, VmObj, CalculateNameObj } from '../../common/interfaces';
 import styled from 'styled-components';
 import { getVmsForSandboxUrl } from '../../../services/ApiCallStrings';
+import useKeyEvents from '../../common/hooks/useKeyEvents';
 import '../../../styles/addNewVm.scss';
 
 const icons = {
@@ -204,6 +205,9 @@ const AddNewVm: React.FC<AddNewVmProps> = ({
     };
 
     const createVm = () => {
+        if (!validateUserInput(vm, loading, vmEstimatedCost, usernameIsValid)) {
+            return;
+        }
         setHasChanged(false);
         setLoading(true);
         setUpdateCache({ ...updateCache, [getVmsForSandboxUrl(sandbox.id)]: true });
@@ -274,6 +278,8 @@ const AddNewVm: React.FC<AddNewVmProps> = ({
         }
         _setFiler(currentFilter);
     };
+
+    useKeyEvents(undefined, createVm, true);
 
     return (
         <Wrapper>
